move api to multigroup support
