+++ controllers/finalizer.go
@@ -8,7 +8,7 @@ package controllers
 import (
 	"context"
 
-	datadoghqv1alpha1 "github.com/DataDog/watermarkpodautoscaler/api/v1alpha1"
+	datadoghqv1alpha1 "github.com/DataDog/watermarkpodautoscaler/api/datadoghq/v1alpha1"
 	"github.com/DataDog/watermarkpodautoscaler/pkg/util"
 	logr "github.com/go-logr/logr"
 )
