+++ controllers/test/suite_test.go
@@ -27,7 +27,7 @@ import (
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
-	datadoghqv1alpha1 "github.com/DataDog/watermarkpodautoscaler/api/v1alpha1"
+	datadoghqv1alpha1 "github.com/DataDog/watermarkpodautoscaler/api/datadoghq/v1alpha1"
 	"github.com/DataDog/watermarkpodautoscaler/controllers"
 	"github.com/DataDog/watermarkpodautoscaler/controllers/test/utils"
 	// +kubebuilder:scaffold:imports
